src/main/java/it/xpug/kata/birthday_greetings/BirthdayService.java
package it.xpug.kata.birthday_greetings;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;
import java.text.ParseException;

import javax.mail.MessagingException;
import javax.mail.internet.AddressException;

public class BirthdayService {

	public void sendGreetings(String fileName, XDate xDate, String smtpHost, int smtpPort) throws IOException, ParseException, AddressException, MessagingException {
		MailSender mailSender = new MailSender(smtpHost, smtpPort);
		sendGreetings(fileName, xDate, mailSender);
	}

	public void sendGreetings(String fileName, XDate xDate, MailSender mailSender) throws IOException, ParseException, MessagingException {
		BufferedReader in = new BufferedReader(new FileReader(fileName));
		String str = "";
		str = in.readLine(); // skip header
		while ((str = in.readLine()) != null) {
			String[] employeeData = str.split(", ");
			Employee employee = new Employee(employeeData[1], employeeData[0], employeeData[2], employeeData[3]);
			if (employee.isBirthday(xDate)) {
				mailSender.sendGreetingsTo(employee);
			}
		}
	}

}
